refactor into blueprints for main sections
