+++ ask_meagain/ask_meagain/blueprints/admin/templates/ResultsPage.html
@@ -16,7 +16,7 @@
         <ul>
         {% for q in questions %}
             <li>
-                <a href="{{url_for('questionResultPage',id = q.id)}}">Question {{q.id}}</a>
+                <a href="{{url_for('admin.questionResultPage',id = q.id)}}">Question {{q.id}}</a>
             </li>
         {% endfor %}
         </ul>
@@ -24,7 +24,7 @@
         <h1> No questions in the Database </h1>
     {% endif %}
 
-    <a href='{{ url_for('menu') }}' target="_parent">
+    <a href='{{ url_for('hello') }}' target="_parent">
         <button class="button" type="button" >Back to Home</button>
     </a>
