+++ ask_meagain/ask_meagain/blueprints/survey/templates/surveyLanding.html
@@ -3,7 +3,7 @@
 {% block title %}Survey Landing - Ask{% endblock %}
 
 {% block content %}
-    <a href="{{ url_for('surveyQuestion') }}" style="text-decoration: none">
+    <a href="{{ url_for('survey.surveyQuestion') }}" style="text-decoration: none">
         <div align="center">
             <h1 style="font-size: 80px">Welcome!</h1>
             <img src="{{ url_for('static', filename='assets/images/Question-mark-blackandwhite.png') }}" class="largeSquareImage">
