+++ ask_meagain/ask_meagain/templates/menu.html
@@ -5,13 +5,13 @@
 {% block content %}
 <div align="center">
     <div style="width: fit-content">
-        <a href="{{url_for('surveyLandingPage')}}" target='_parent'>
+        <a href="{{url_for('survey.surveyLandingPage')}}" target='_parent'>
             <div class="rectangle" style="float:left;">
                 <h1>Take Survey</h1>
                 <img src="{{url_for('static', filename='assets/images/220px-Pictogram_voting_question.svg.png')}}" class="squareImage">
             </div>
         </a>
-        <a href='{{ url_for('surveyLandingPage') }}' target="_parent">
+        <a href='{{ url_for('survey.surveyLandingPage') }}' target="_parent">
             <div class="rectangle" style = "float:right;">
                 <h1>Admin Console</h1>
                 <img src="{{url_for('static', filename='assets/images/images.png')}}" class="squareImage">
