+++ ask_meagain/ask_meagain/templates/thankYou.html
@@ -3,7 +3,7 @@
 {% block title %}Thank You - Ask{% endblock %}
 
 {% block content %}
-<a href="{{url_for('menu')}}" style="text-decoration: none">
+<a href="{{url_for('hello')}}" style="text-decoration: none">
 <div align="center">
     <h1 style="font-size: 80px">Thank You!</h1>
     <img src="{{url_for('static', filename='assets/images/Question-mark-blackandwhite.png')}}" class="largeSquareImage">
